Refactor Mono thread attach & safe unboxing patterns
Refactored thread attachment in mono with ensure_thread_attached_impl() and added a generic invoke_compiled_method_impl template for type-safe Mono method invocation. Updated all invoke usages to check for null MonoObject before unboxing, preventing null dereferences. Changed player::get_player_id() to return int64_t and use correct Mono method. Applied safe unboxing and null checks to food and Vector3-returning methods. Minor formatting and namespace adjustments for consistency.

+++ src/unity/player.hpp
@@ -37,7 +37,7 @@ namespace big
 		float get_eitr_regen();
 		float get_max_adrenalin();
 		float get_adrenalin();
-		int get_player_id();
+		int64_t get_player_id();
 		bool is_player() override;
 		skills get_skills();
 		std::string get_player_name();
